Fixed a bug in update_registry_server workflow
For some reason, the somewhat aged ccv3 client that is
in our vendor bundle is unable to `UpdateApplication`
if given an application object retrieved from the API.

The error presented is "Unknown field(s): relationships"
and is being generated on the capi side of the equation.

We've worked around this by creating a temporary copy of
the app object without its relationships collection and
passing THAT temp object to the update call.

It appears that this likely also affects update_config_server
workflow, and we should double-check that.